Put Material 3 usage in a Composition Local
This lets us not have to pass `useMaterial3` around all over the place.
Instead, we set it on the top (screen) level, and then it defaults to
whichever screen was set further down the line.

Before, we had to do something like this:

```
M3SatsScreen {
  SatsSurface(useMaterial3 = true) {
    SatsBanner("Banner text", useMaterial3 = true)
  }
}
```

Now, it should be enough just to use an `M3SatsScreen`, and the
components will respect the Composition Local:

```
M3SatsScreen {
  SatsSurface {
    SatsBanner("Banner text")
  }
}
```