Restrict the SatsBanner API surface
The banner should always use the same colors, as per the Design System.
Thus, it doesn't make sense to allow that as part of the API surface.

Similarly, the optional action should always be represented by a
Clean Button, so allowing a slot for the button is also senseless. Thus,
a new `SatsBannerAction` has been introduced in its stead.