Fix colors for `SatsCard` when in Material 3
This switches `SatsCard` from using the `ElevatedCard` component to just
creating our own version using a `SatsSurface` when in a Material 3
context. This is because `ElevatedCard` also applies a tonal elevation
to the surface based on how much elevation we have, which looks very
wrong in the app. Instead, we always want a regular surface color with
some shadows around the card, like we have in Material 2 already.