Create new Form Text Field components
These mirror the iOS form text fields that we have in the design system
and that Apple seems to use in settings and stuff. I wasn't able to find
them anywhere in the Human Interface Guidelines, however, so I just
YOLO'd a little bit on the spacing based on what was in Figma. I also
made sure the touch areas were large enough for Material guidelines,
which is slightly larger than what's expected on iOS.