Add form input field with custom content
This field looks identical to the single-line form input fields, except
it's not backed by a text field. Instead, it takes any kind of content.
This makes it ideal form fields that allow the user to select from a
list or a drop-down menu, or for controls like switches/checkboxes.

Note that these form fields are always single-line.