Create new date/time form field component
This form field component shows a selected date and time in two separate
boxes in the content position of the form input field. It provides two
formatting callbacks (one for the date and another for the time) and two
callbacks for when the fields are clicked. It does not deal with opening
any date or time picker – that's part is left to the caller.