Add optional trailing text to form text field
Adds a new trailing text property to form text fields, which will be
rendered at after the actual text field. This option only makes sense in
single-line text fields, and will only be rendered in those cases. This
property is meant to be used for things like units of measurement, such
as “minutes” or “kilometers”.