Add input transformations to form text fields
Input transformations allow restricting the input provided to the form
field by providing a lambda that gets access to the original value and
its changes. The changes can either be allowed through (meaning there's
no transformation), reverted in its entirety, or be mapped over and
manipulated one by one.

The sample app uses an input transformation to apply a digits-only
filter to the form's Duration field.