Add keyboard options + actions to form text fields
Keyboard options allow customizing how the on-screen keyboard looks.
This includes things like which type of keyboard should be shown
(numeric, e-mail address, URI, …), and what kind of capitalization
should be used (none, words, sentences, …). It also allows setting what
the “action” button should be.

Keyboard actions can be used to react to action button clicks, such as
calling a function to send the form data to the back-end when the user
clicks on the Send keyboard action, given that the Send action was set
as the action button through the keyboard options.